chore(backward): moved backward_compat files inside tfhe-rs modules
