--- tfhe/src/backward_compatibility/mod.rs
@@ -1,3 +1,0 @@
-pub mod core_crypto;
-#[cfg(feature = "shortint")]
-pub mod shortint;
