+++ tfhe/src/core_crypto/commons/math/random/mod.rs
@@ -13,7 +13,7 @@
 //! implementations of this trait, for different distributions. Note, though, that instead of
 //! using the [`RandomGenerable`] methods, you should use the various methods exposed by
 //! [`RandomGenerator`] instead.
-use crate::backward_compatibility::core_crypto::commons::math::random::DynamicDistributionVersions;
+use crate::core_crypto::backward_compatibility::commons::math::random::DynamicDistributionVersions;
 use crate::core_crypto::commons::dispersion::{DispersionParameter, StandardDev, Variance};
 use crate::core_crypto::commons::numeric::{FloatingPoint, UnsignedInteger};
 use std::ops::Bound;
