+++ tfhe/src/core_crypto/commons/math/random/t_uniform.rs
@@ -1,4 +1,4 @@
-use crate::backward_compatibility::core_crypto::commons::math::random::TUniformVersions;
+use crate::core_crypto::backward_compatibility::commons::math::random::TUniformVersions;
 
 use super::*;
 use serde::{Deserialize, Serialize};
