+++ tfhe/src/core_crypto/entities/lwe_ciphertext.rs
@@ -2,8 +2,8 @@
 
 use tfhe_versionable::Versionize;
 
-use crate::backward_compatibility::core_crypto::entities::lwe_ciphertext::LweCiphertextVersions;
 use crate::conformance::ParameterSetConformant;
+use crate::core_crypto::backward_compatibility::entities::lwe_ciphertext::LweCiphertextVersions;
 use crate::core_crypto::commons::parameters::*;
 use crate::core_crypto::commons::traits::*;
 use crate::core_crypto::prelude::misc::check_encrypted_content_respects_mod;
