+++ tfhe/src/shortint/parameters/mod.rs
@@ -5,7 +5,6 @@
 //! homomorphic evaluation of integer circuits as well as a list of secure cryptographic parameter
 //! sets.
 
-use crate::backward_compatibility::shortint::parameters::*;
 use crate::conformance::ListSizeConstraint;
 pub use crate::core_crypto::commons::dispersion::StandardDev;
 pub use crate::core_crypto::commons::parameters::{
@@ -13,6 +12,7 @@ pub use crate::core_crypto::commons::parameters::{
     DynamicDistribution, GlweDimension, LweBskGroupingFactor, LweDimension, PolynomialSize,
 };
 use crate::core_crypto::prelude::{LweCiphertextListParameters, LweCiphertextParameters};
+use crate::shortint::backward_compatibility::parameters::*;
 use serde::{Deserialize, Serialize};
 
 use tfhe_versionable::Versionize;
